pyproject.toml
[project]
name = "multiagent-eval"
version = "1.2.1"
description = "Enterprise-grade reliability framework for AI agents."
readme = "README.md"
authors = [
    { name = "Najeed Khan", email = "ai.eval.harness.contact@gmail.com" }
]
license = "Apache-2.0"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Intended Audience :: Developers",
]
dependencies = [
    "aiohttp>=3.8.0",
    "Flask>=2.0.0",
    "flask-cors>=3.0.0",
    "requests>=2.0.0",
    "jsonschema>=4.0.0",
    "PyYAML>=6.0",
    "sentence-transformers",
    "numpy",
    "datasets",
    "PyJWT>=2.0.0",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["eval_runner", "dataproc_engine", "industries"]

[tool.setuptools.package-data]
"*" = ["*.json", "*.jsonl", "*.csv", "*.md", "*.yaml", "*.yml"]

[project.scripts]
multiagent-eval = "eval_runner.cli:main"

[project.entry-points."eval_runner.plugins"]
coverage = "eval_runner.coverage_plugin:CoveragePlugin"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra"
python_files = ["test_*.py"]